Add/Organize tests for ctx
